requirements.txt
fastcore
fastprogress
openai
arize-phoenix[evals]